requirements.txt
# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.0
requests==2.31.0
webdriver-manager==4.0.1
lxml==4.9.3

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# NLP & Analysis
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2

# AI API - NVIDIA NIM (resume rewriting + document classification)
openai>=1.0.0

# PDF Generation
reportlab==4.0.7
python-docx==1.1.0
pdfminer.six>=20221105

# Scheduling
apscheduler==3.10.4

# CLI
click==8.1.7
rich==13.7.0

# Web Framework (Phase 8 - Optional)
flask==3.0.0
flask-cors>=4.0.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0
responses==0.24.1
